requirements.txt
# requirements.txt

fastapi
python-dotenv
uvicorn[standard]
tqdm
python-multipart
setuptools
pybind11
numpy
faiss-cpu